添加敏感词接口
添加敏感词接口：1.批量导入2.增3.删4.改

+++ TerraMours/TerraMours/Framework/Infrastructure/Contracts/SystemModels/BaseEntity.cs
@@ -20,7 +20,7 @@
         /// <summary>
         /// 创建人id
         /// </summary>
-        public int? CreateID { get; set; }
+        public long? CreateID { get; set; }
         /// <summary>
         /// 创建人名称
         /// </summary>
@@ -28,7 +28,7 @@
         /// <summary>
         /// 修改人id
         /// </summary>
-        public int? ModifyID { get; set; }
+        public long? ModifyID { get; set; }
         /// <summary>
         /// 修改人名称
         /// </summary>
